+++ src/compas/scene/__init__.py
@@ -9,7 +9,6 @@ from __future__ import absolute_import
 from __future__ import division
 
 from .exceptions import SceneObjectNotRegisteredError
-from .exceptions import NoSceneObjectContextError
 from .sceneobject import SceneObject
 from .meshobject import MeshObject
 from .graphobject import GraphObject
@@ -26,9 +25,23 @@ from .scene import Scene
 from .scene import SceneObjectNode
 from .scene import SceneTree
 
+from compas.plugins import plugin
+from compas.geometry import Geometry
+from compas.datastructures import Mesh
+from compas.datastructures import Graph
+from compas.datastructures import VolMesh
+
+
+@plugin(category="factories", pluggable_name="register_scene_objects")
+def register_scene_objects_base():
+    register(Geometry, GeometryObject, context=None)
+    register(Mesh, MeshObject, context=None)
+    register(Graph, GraphObject, context=None)
+    register(VolMesh, VolMeshObject, context=None)
+
+
 __all__ = [
     "SceneObjectNotRegisteredError",
-    "NoSceneObjectContextError",
     "SceneObject",
     "MeshObject",
     "GraphObject",